Added riccati submodule
