+++ sympy/solvers/ode/ode.py
@@ -7097,4 +7097,5 @@ def _nonlinear_3eq_order1_type5(x, y, z, t, eq):
 #This import is written at the bottom to avoid circular imports.
 from .single import (NthAlgebraic, Factorable, FirstLinear, AlmostLinear,
         Bernoulli, SingleODEProblem, SingleODESolver, RiccatiSpecial,
-        SecondNonlinearAutonomousConserved, RationalRiccati)
+        SecondNonlinearAutonomousConserved)
+from .riccati import RationalRiccati
